requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py38_0
astroid=2.3.3=py38_0
attrs=19.3.0=py_0
backcall=0.1.0=py38_0
blas=1.0=mkl
bleach=3.1.4=py_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py38_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38_0
importlib-metadata=1.6.0=py38_0
importlib_metadata=1.6.0=0
intel-openmp=2019.4=233
ipykernel=5.1.4=py38h39e3cac_0
ipython=7.13.0=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
isort=4.3.21=py38_0
jedi=0.17.0=py38_0
jinja2=2.11.2=py_0
jsonschema=3.2.0=py38_0
jupyter_client=6.1.3=py_0
jupyter_core=4.6.3=py38_0
krb5=1.16.4=hddcf347_0
lazy-object-proxy=1.4.3=py38h1de35cc_0
libcxx=10.0.0=1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h0a44026_6
libgfortran=3.0.1=h93005f0_2
libpq=11.2=h051b688_0
libsodium=1.0.16=h3efe00b_0
markupsafe=1.1.1=py38h1de35cc_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h1de35cc_1000
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.0.15=py38h5e564d8_0
mkl_random=1.1.1=py38h959d312_0
nbconvert=5.6.1=py38_0
nbformat=5.0.6=py_0
ncurses=6.2=h0a44026_1
notebook=6.0.3=py38_0
numpy=1.18.1=py38h7241aed_0
numpy-base=1.18.1=py38h3304bdc_1
openssl=1.1.1g=h1de35cc_0
pandas=1.0.3=py38h6c726b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pip=20.0.2=py38_3
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
psycopg2=2.8.4=py38ha12b0ac_0
ptyprocess=0.6.0=py38_0
pygments=2.6.1=py_0
pylint=2.4.4=py38_0
pyrsistent=0.16.0=py38h1de35cc_0
python=3.8.1=h359304d_1
python-dateutil=2.8.1=py_0
python-wget=3.2=py_0
pytz=2020.1=py_0
pyzmq=18.1.1=py38h0a44026_0
readline=7.0=h1de35cc_5
send2trash=1.5.0=py38_0
setuptools=46.4.0=py38_0
six=1.14.0=py38_0
sqlalchemy=1.3.17=py38haf1e3a3_0
sqlite=3.31.1=h5c1f38d_1
terminado=0.8.3=py38_0
testpath=0.4.4=py_0
tk=8.6.8=ha441bb4_0
tornado=6.0.4=py38h1de35cc_1
traitlets=4.3.3=py38_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py38_1
wheel=0.34.2=py38_0
wrapt=1.12.1=py38h1de35cc_1
xz=5.2.5=h1de35cc_0
zeromq=4.3.1=h0a44026_3
zipp=3.1.0=py_0
zlib=1.2.11=h1de35cc_3
